server/requirements.txt
flask==2.2.2
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
marshmallow==3.19.0
black==23.1.0
isort==5.12.0